third_party/WebKit/LayoutTests/imported/wpt/html/webappapis/scripting/processing-model-2/runtime-error-data-url-expected.txt

CONSOLE ERROR: line 1: Uncaught ReferenceError: undefined_variable is not defined
This is a testharness.js-based test.
PASS window.onerror - runtime error in <script src=data:...> 
PASS window.onerror - runtime error in <script src=data:...> (column) 
Harness: the test ran to completion.
